Exclude nested properties from the reference resolution
