+++ CHANGELOG.md
@@ -21,6 +21,9 @@
 - Adjusted `log4j2` live template [#781](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/781)
 - Adjusted `slf4j` live template [#782](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/782)
 
+### `OCC` enhancements
+- Exclude nested properties from the reference resolution [#799](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/799)
+
 ### Other
 - Allow modification of the non-project root files [#794](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/794)
 - Adjust type system merge for `Product` type (it extends both `GenericItem` and `LocalizableItem`) [#785](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/785)
