+++ src/com/intellij/idea/plugin/hybris/system/bean/codeInsight/completion/BSCompletionService.kt
@@ -19,15 +19,49 @@
 package com.intellij.idea.plugin.hybris.system.bean.codeInsight.completion
 
 import com.intellij.codeInsight.lookup.LookupElement
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.system.bean.codeInsight.lookup.BSLookupElementFactory
+import com.intellij.idea.plugin.hybris.system.bean.meta.BSMetaModelAccess
 import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSGlobalMetaBean
+import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSGlobalMetaEnum
 import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSMetaType
+import com.intellij.openapi.components.Service
 import com.intellij.openapi.project.Project
 
-interface BSCompletionService {
+@Service(Service.Level.PROJECT)
+class BSCompletionService(private val project: Project) {
 
-    fun getCompletions(meta: BSGlobalMetaBean): List<LookupElement>
+    fun getCompletions(meta: BSGlobalMetaBean): List<LookupElement> {
+        val properties = meta.allProperties.values
+            .mapNotNull { BSLookupElementFactory.build(it) }
+        val levelMappings = HybrisConstants.OCC_DEFAULT_LEVEL_MAPPINGS
+            .map { BSLookupElementFactory.buildLevelMapping(it) }
+        return properties + levelMappings
+    }
 
-    fun getCompletions(vararg types: BSMetaType): List<LookupElement>
+    fun getCompletions(vararg types: BSMetaType) = with(BSMetaModelAccess.getInstance(project)) {
+        types
+            .map { metaType ->
+                when (metaType) {
+                    BSMetaType.META_ENUM -> this
+                        .getAll<BSGlobalMetaEnum>(metaType)
+                        .mapNotNull { BSLookupElementFactory.build(it) }
+
+                    BSMetaType.META_BEAN -> this
+                        .getAll<BSGlobalMetaBean>(metaType)
+                        .mapNotNull { BSLookupElementFactory.build(it, metaType) }
+
+                    BSMetaType.META_WS_BEAN -> this
+                        .getAll<BSGlobalMetaBean>(metaType)
+                        .mapNotNull { BSLookupElementFactory.build(it, metaType) }
+
+                    BSMetaType.META_EVENT -> this
+                        .getAll<BSGlobalMetaBean>(metaType)
+                        .mapNotNull { BSLookupElementFactory.build(it, metaType) }
+                }
+            }
+            .flatten()
+    }
 
     companion object {
         fun getInstance(project: Project): BSCompletionService = project.getService(BSCompletionService::class.java)
