+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/BSMetaModelCollector.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -17,14 +17,56 @@
  */
 package com.intellij.idea.plugin.hybris.system.bean.meta
 
+import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
+import com.intellij.idea.plugin.hybris.system.bean.BSUtils
+import com.intellij.idea.plugin.hybris.system.bean.model.Beans
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.progress.ProgressManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
+import com.intellij.psi.search.ProjectScope
+import com.intellij.psi.stubs.StubIndex
+import com.intellij.psi.xml.XmlFile
+import com.intellij.util.Processor
+import com.intellij.util.xml.DomManager
+import com.intellij.util.xml.stubs.index.DomElementClassIndex
+import java.util.*
 
-interface BSMetaModelCollector {
+@Service(Service.Level.PROJECT)
+class BSMetaModelCollector(private val myProject: Project) {
 
     companion object {
         fun getInstance(project: Project): BSMetaModelCollector = project.getService(BSMetaModelCollector::class.java)
     }
 
-    fun collectDependencies(): Set<PsiFile>
+    private val myDomManager: DomManager = DomManager.getDomManager(myProject)
+
+    fun collectDependencies(): Set<PsiFile> {
+        val files = HashSet<PsiFile>()
+
+        ProgressManager.getInstance().progressIndicator.text2 = HybrisI18NBundleUtils.message("hybris.bs.access.progress.subTitle.collectingDependencies")
+
+        StubIndex.getInstance().processElements(
+            DomElementClassIndex.KEY,
+            Beans::class.java.name,
+            myProject,
+            ProjectScope.getAllScope(myProject),
+            PsiFile::class.java,
+            object : Processor<PsiFile> {
+                override fun process(psiFile: PsiFile): Boolean {
+                    psiFile.virtualFile ?: return true
+                    // cannot process a file without a module
+                    BSUtils.getModuleForFile(psiFile) ?: return true
+                    myDomManager.getFileElement(psiFile as XmlFile, Beans::class.java) ?: return true
+
+                    files.add(psiFile)
+                    return true
+                }
+            }
+        )
+
+        ProgressManager.getInstance().progressIndicator.text2 = HybrisI18NBundleUtils.message("hybris.bs.access.progress.subTitle.collectedDependencies", files.size)
+
+        return Collections.unmodifiableSet(files)
+    }
 }
