+++ src/com/intellij/idea/plugin/hybris/system/bean/psi/OccPropertyMapping.kt
@@ -1,0 +1,29 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.intellij.idea.plugin.hybris.system.bean.psi
+
+import com.intellij.openapi.util.TextRange
+
+data class OccPropertyMapping(
+    private val startPosition: Int,
+    val value: String,
+    val textRange: TextRange = TextRange.from(startPosition, value.length),
+    val children: MutableList<OccPropertyMapping> = mutableListOf(),
+    var parent: OccPropertyMapping? = null
+)
