+++ src/com/intellij/idea/plugin/hybris/system/bean/psi/reference/OccBSBeanPropertyReference.kt
@@ -22,8 +22,8 @@ import com.intellij.codeInsight.highlighting.HighlightedReference
 import com.intellij.idea.plugin.hybris.psi.util.PsiUtils
 import com.intellij.idea.plugin.hybris.system.bean.codeInsight.completion.BSCompletionService
 import com.intellij.idea.plugin.hybris.system.bean.meta.model.BSGlobalMetaBean
+import com.intellij.idea.plugin.hybris.system.bean.psi.OccPropertyMapping
 import com.intellij.idea.plugin.hybris.system.bean.psi.reference.result.BeanPropertyResolveResult
-import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiPolyVariantReference
 import com.intellij.psi.PsiReferenceBase
@@ -32,8 +32,8 @@ import com.intellij.psi.ResolveResult
 class OccBSBeanPropertyReference(
     private val meta: BSGlobalMetaBean,
     element: PsiElement,
-    range: TextRange
-) : PsiReferenceBase.Poly<PsiElement>(element, range, false), PsiPolyVariantReference, HighlightedReference {
+    private val mapping: OccPropertyMapping
+) : PsiReferenceBase.Poly<PsiElement>(element, mapping.textRange, false), PsiPolyVariantReference, HighlightedReference {
 
     override fun getVariants() = BSCompletionService.getInstance(element.project)
         .getCompletions(meta)
